Refactor code to categorise them into packages
