+++ src/main/java/duke/exception/DukeException.java
@@ -1,5 +1,7 @@
+package duke.exception;
+
 /**
- * Provides Duke class an Exception.
+ * Provides duke.duke class an Exception.
  */
 public class DukeException extends Exception {
     public DukeException(String errMsg) {
