+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,10 @@
+package duke.storage;
+
+import duke.exception.DukeException;
+import duke.task.TaskList;
+import duke.parser.Parser;
+import duke.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -35,9 +42,9 @@ public class Storage {
                 f.getParentFile().mkdirs();
                 f.createNewFile();
             } catch (IOException ioe) {
-                throw new DukeException("Fake Duke can't create the file.");
+                throw new DukeException("Fake duke.duke can't create the file.");
             }
-            throw new DukeException(String.format("Fake Duke can't find the file. I will create the file (%s) :D", this.filePath));
+            throw new DukeException(String.format("Fake duke.duke can't find the file. I will create the file (%s) :D", this.filePath));
         }
         return tasks;
     }
