+++ src/main/java/duke/task/Deadline.java
@@ -1,10 +1,13 @@
-import java.time.LocalDate;
+package duke.task;
+
+import duke.exception.DukeException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 /**
- * Represents a Deadline, which is a type of Task that has to be done before s specific date/time.
+ * Represents a duke.Deadline, which is a type of duke.Task that has to be done before s specific date/time.
  */
 public class Deadline extends Task {
     protected LocalDateTime deadline;
@@ -14,16 +17,16 @@ public class Deadline extends Task {
         try {
             this.deadline = this.parseDatetime(description.split(" /by ")[1]);
         } catch (ArrayIndexOutOfBoundsException e) {
-            throw new DukeException("☹ I'm sorry, but Fake Duke doesn't know what that means :-(");
+            throw new DukeException("☹ I'm sorry, but Fake duke.duke doesn't know what that means :-(");
         } catch (DateTimeParseException dtpe) {
             throw new DukeException("Invalid datetime format. Please use yyyy-mm-dd HH:mm (E.g. 2019-10-15 18:00).");
         }
     }
 
     /**
-     * Returns the String representation of a Deadline.
+     * Returns the String representation of a duke.Deadline.
      *
-     * @return  String representation of a Deadline in this format: [type][<status>] <description> (by: <deadline>).
+     * @return  String representation of a duke.Deadline in this format: [type][<status>] <description> (by: <deadline>).
      */
     @Override
     public String toString() {
