+++ src/main/java/duke/task/Event.java
@@ -1,9 +1,13 @@
+package duke.task;
+
+import duke.exception.DukeException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 /**
- * Represents an Event, which is a type of Task that starts at a specific date/time and ends at a specific date/time.
+ * Represents an duke.Event, which is a type of duke.Task that starts at a specific date/time and ends at a specific date/time.
  */
 public class Event extends Task {
     protected LocalDateTime startDatetime;
@@ -16,16 +20,16 @@ public class Event extends Task {
             this.startDatetime = parseDatetime(datetimes.split(" /to ")[0]);
             this.endDatetime = parseDatetime(datetimes.split(" /to ")[1]);
         } catch (ArrayIndexOutOfBoundsException e) {
-            throw new DukeException("☹ I'm sorry, but Fake Duke doesn't know what that means :-(");
+            throw new DukeException("☹ I'm sorry, but Fake duke.duke doesn't know what that means :-(");
         } catch (DateTimeParseException dtpe) {
             throw new DukeException("Invalid datetime format. Please use yyyy-mm-dd HH:mm (E.g. 2019-10-15 18:00).");
         }
     }
 
     /**
-     * Returns the String representation of an Event.
+     * Returns the String representation of an duke.Event.
      *
-     * @return  String representation of an Event in this format: [E][<status>] <description> (from: <start date/time>
+     * @return  String representation of an duke.Event in this format: [E][<status>] <description> (from: <start date/time>
      * to: <end date/time>).
      */
     @Override
