+++ text-ui-test/data/tasks.txt
@@ -1,0 +1,7 @@
+T ~ 0 ~ read book
+T ~ 0 ~ read book
+T ~ 0 ~ read book
+D ~ 0 ~ return book ~ 2019-09-02 12:22
+D ~ 0 ~ return book ~ 2019-09-02 12:22
+D ~ 0 ~ return book ~ 2019-09-02 12:22
+D ~ 0 ~ return book ~ 2019-09-02 12:22
